[REFACTOR/#123] 구매 모듈 설정
