+++ feature/buy/build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 }
 
 android {
-    namespace = "co.orange.feature.buy"
+    namespace = "co.orange.buy"
     compileSdk = Constants.compileSdk
 
     defaultConfig {
@@ -16,6 +16,12 @@ android {
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles("consumer-rules.pro")
+
+        buildConfigField(
+            "String",
+            "IAMPORT_CODE",
+            gradleLocalProperties(rootDir).getProperty("iamport.code"),
+        )
 
         buildConfigField(
             "String",
