+++ feature/buy/src/main/res/layout/activity_buy_finished.xml
