+++ presentation/src/main/java/co/orange/presentation/detail/OptionBottomSheet.kt
@@ -11,7 +11,7 @@ import co.orange.core.R
 import co.orange.core.base.BaseBottomSheet
 import co.orange.core.extension.setOnSingleClickListener
 import co.orange.core.extension.setOverThousand
-import co.orange.presentation.buy.progress.BuyProgressActivity
+import com.kkkk.buy.progress.BuyProgressActivity
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
@@ -56,7 +56,7 @@ class OptionBottomSheet :
 
     private fun initPurchaseBtnListener() {
         binding.btnPurchase.setOnSingleClickListener {
-            BuyProgressActivity.createIntent(
+            com.kkkk.buy.progress.BuyProgressActivity.createIntent(
                 requireContext(),
                 viewModel.productId,
                 viewModel.selectedOptionList,
